requirements.txt
python-telegram-bot==20.7
openai
apscheduler
mysql-connector-python
sqlalchemy
python-dotenv
python-telegram-bot[job-queue]==20.7
pymysql
matplotlib
pillow
numpy
seaborn